moved views to view.py, moved models to model.py, and moved passwordmanager to pm.py
